requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
pdf2docx
pillow
PyMuPDF
